# app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persona API",
    "description": "Persona service",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Is the service healthy",
        "description": "Is the service healthy",
        "responses": {
          "200": {
            "description": "The service is running",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    },
    "/persona": {
      "post": {
        "summary": "Create a persona",
        "description": "Create a persona",
        "security": [
          {
            "authHeader": []
          }
        ],
        "requestBody": {
          "description": "The persona data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "The unique identifier for the persona"
                  },
                  "firstname": {
                    "type": "string",
                    "description": "The first name of the persona"
                  },
                  "lastname": {
                    "type": "string",
                    "description": "The last name of the persona"
                  },
                  "interests": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "An array of interests for the persona"
                  },
                  "latitude": {
                    "type": "number",
                    "format": "float",
                    "description": "The latitude coordinate of the persona's location"
                  },
                  "longitude": {
                    "type": "number",
                    "format": "float",
                    "description": "The longitude coordinate of the persona's location"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Persona created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Persona"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input.  Missing required fields.  Id is not unique."
          },
          "403": {
            "description": "Invalid authorization."
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    },
    "/persona/{id}": {
      "get": {
        "summary": "Get persona by Id",
        "description": "Get persona by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "authHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "Persona retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Persona"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input.  Missing required id."
          },
          "403": {
            "description": "Invalid authorization."
          },
          "404": {
            "description": "Persona with ID not found."
          },
          "5XX": {
            "description": "Unexpected error."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "authHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    },
    "schemas": {
      "Persona": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Externally defined unique id string for the Persona"
          },
          "firstname": {
            "type": "string",
            "description": "The first name for the persona"
          },
          "lastname": {
            "type": "string",
            "description": "The last name for the persona"
          },
          "intrests": {
            "description": "An array of interests for the persona",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "latitude": {
            "type": "number",
            "description": "The latitude of the persona's location"
          },
          "longitude": {
            "type": "number",
            "description": "The longitude of the persona's location"
          },
          "city": {
            "type": "string",
            "description": "The city for the lat/long according by the National Weather Service"
          },
          "state": {
            "type": "string",
            "description": "The state for the lat/long according by the National Weather Service"
          },
          "current_tempurature": {
            "type": "number",
            "description": "The current tempurature for the lat/long according by the National Weather Service"
          }
        }
      }
    }
  },
  "tags": []
}